Some light structural refactoring
