+++ lib/core/instance.js
@@ -2,17 +2,11 @@
 
 var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
 
-function _toConsumableArray(arr) { if (Array.isArray(arr)) { for (var i = 0, arr2 = Array(arr.length); i < arr.length; i++) { arr2[i] = arr[i]; } return arr2; } else { return Array.from(arr); } }
-
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
-require('colors');
 var winston = require('winston'),
-    fs = require('fs');
-
-// @TODO: Can we profile when we run two concurrent tasks.
-//  bolt A B
-//  profile when both A and B finish.
+    fs = require('fs'),
+    BoltTask = require('./task');
 
 var BoltInstance = function () {
   function BoltInstance() {
@@ -229,65 +223,5 @@ var BoltInstance = function () {
 
   return BoltInstance;
 }();
-//
 
-
-var BoltTask = function () {
-  function BoltTask(parent, opts) {
-    _classCallCheck(this, BoltTask);
-
-    Object.assign(this, opts);
-    this.parent = parent;
-    if (opts.deps.length > 0) {
-      this.deps = [];
-      var _iteratorNormalCompletion6 = true;
-      var _didIteratorError6 = false;
-      var _iteratorError6 = undefined;
-
-      try {
-        for (var _iterator6 = opts.deps[Symbol.iterator](), _step6; !(_iteratorNormalCompletion6 = (_step6 = _iterator6.next()).done); _iteratorNormalCompletion6 = true) {
-          var dep = _step6.value;
-
-          this.deps.push(require(dep));
-        }
-      } catch (err) {
-        _didIteratorError6 = true;
-        _iteratorError6 = err;
-      } finally {
-        try {
-          if (!_iteratorNormalCompletion6 && _iterator6.return) {
-            _iterator6.return();
-          }
-        } finally {
-          if (_didIteratorError6) {
-            throw _iteratorError6;
-          }
-        }
-      }
-    }
-  }
-
-  _createClass(BoltTask, [{
-    key: 'run',
-    value: function run(env) {
-      var _this2 = this;
-
-      return new Promise(function (resolve, reject) {
-        winston.info('Running ' + _this2.name);
-        winston.profile(_this2.name);
-        if (_this2.func && typeof _this2.func === 'function') _this2.func.apply(_this2, _toConsumableArray(_this2.deps).concat([{
-          env: env,
-          config: _this2.parent.config,
-          resolve: resolve,
-          reject: reject,
-          run: _this2.parent.runTask.bind(_this2.parent)
-        }]));
-      });
-    }
-  }]);
-
-  return BoltTask;
-}();
-
-exports.BoltInstance = BoltInstance;
-exports.BoltTask = BoltTask;
+module.exports = BoltInstance;
