+++ src/lib/core/instance.js
@@ -1,6 +1,7 @@
-require('colors');
 const winston  = require('winston'),
-  fs           = require('fs');
+  fs           = require('fs'),
+  BoltTask     = require('./task');
+
 
 class BoltInstance {
   constructor() {
@@ -106,32 +107,5 @@ class BoltInstance {
     }
   }
 }
-//
-class BoltTask {
-  constructor(parent, opts) {
-    Object.assign(this, opts);
-    this.parent = parent;
-    if (opts.deps.length > 0) {
-      this.deps = [];
-      for (const dep of opts.deps)
-        this.deps.push(require(dep));
-    }
-  }
-  run(env) {
-    return new Promise((resolve, reject) => {
-      winston.info(`Running ${this.name}`);
-      winston.profile(this.name);
-      if (this.func && typeof this.func === 'function')
-        this.func(...this.deps, {
-          env: env,
-          config: this.parent.config,
-          resolve: resolve,
-          reject: reject,
-          run: this.parent.runTask.bind(this.parent)
-        });
-    });
-  }
-}
 
-exports.BoltInstance = BoltInstance;
-exports.BoltTask = BoltTask;
+module.exports = BoltInstance;
